+++ translate_ts2desktop.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-DESKTOP_TEMP_FILE=deepin-graphics-driver-manager.desktop.tmp
-DESKTOP_SOURCE_FILE=deepin-graphics-driver-manager.desktop
-DESKTOP_DEST_FILE=deepin-graphics-driver-manager.desktop
+DESKTOP_TEMP_FILE=gxde-graphics-driver-manager.desktop.tmp
+DESKTOP_SOURCE_FILE=gxde-graphics-driver-manager.desktop
+DESKTOP_DEST_FILE=gxde-graphics-driver-manager.desktop
 DESKTOP_TS_DIR=translations/desktop/
 
 /usr/bin/deepin-desktop-ts-convert ts2desktop $DESKTOP_SOURCE_FILE $DESKTOP_TS_DIR $DESKTOP_TEMP_FILE